+++ prettier.config.js
@@ -1,10 +1,10 @@
 module.exports = {
-  printWidth: 80,
-  trailingComma: 'all',
-  tabWidth: 4,
-  semi: true,
-  singleQuote: true,
-  arrowParens: 'avoid',
-  bracketSameLine: true,
-  endOfLine: 'lf',
+    printWidth: 80,
+    trailingComma: 'all',
+    tabWidth: 4,
+    semi: true,
+    singleQuote: true,
+    arrowParens: 'avoid',
+    bracketSameLine: true,
+    endOfLine: 'lf',
 };
