+++ src/types/Character/helpers.ts
@@ -1,11 +1,16 @@
 import { JsonValue } from '../../utils';
+import { Bonus } from './Bonus';
 import { Class } from './Class';
 import { Feat } from './Feats';
 import Race from './Race';
+import Stats, { StatSet } from './Stats';
 
 const defaultRace = 'human.cha';
 const Races = Race.import();
-function checkRace(raceName: string): Race {
+function checkRace(raceName: string | Race): Race {
+    if (raceName instanceof Race) {
+        return raceName;
+    }
     const r = Races[raceName || defaultRace];
     if (!r) {
         throw new Error(`Unknown race key: ${raceName}`);
@@ -72,5 +77,29 @@ function exportFeatchChoices(...raw: FeatChoice[]): FeatChoiceExport[] {
     }));
 }
 
+function buildStats({
+    base,
+    race,
+    auras,
+}: {
+    base: StatSet;
+    race: Race;
+    auras: Bonus;
+}): Stats {
+    return new Stats({
+        base,
+        racial: race.statMods,
+        // TODO ? enhance
+        misc: auras.stats.values,
+        // TODO ? temp
+    });
+}
+
 export type { FeatChoice, FeatChoiceExport };
-export { checkRace, checkClass, parseFeatChoices, exportFeatchChoices };
+export {
+    checkRace,
+    checkClass,
+    parseFeatChoices,
+    exportFeatchChoices,
+    buildStats,
+};
